+++ lerna.json
@@ -1,6 +1,6 @@
 {
   "npmClient": "npm",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "bootstrap": {
     "hoist": true
   },
