+++ packages/controls/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oasis-engine/controls",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "scripts": {
     "b:types": "tsc",
@@ -15,6 +15,6 @@
     "types/**/*"
   ],
   "dependencies": {
-    "oasis-engine": "0.2.0"
+    "oasis-engine": "0.2.1"
   }
 }
