+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oasis-engine/core",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "main": "dist/main.js",
   "module": "dist/module.js",
@@ -13,9 +13,9 @@
     "types/**/*"
   ],
   "dependencies": {
-    "@oasis-engine/math": "0.2.0"
+    "@oasis-engine/math": "0.2.1"
   },
   "devDependencies": {
-    "@oasis-engine/design": "0.2.0"
+    "@oasis-engine/design": "0.2.1"
   }
 }
