+++ packages/core/src/material/index.ts
@@ -1,8 +1,10 @@
-export { BlinnPhongMaterial } from "./BlinnPhongMaterial";
 export { AlphaMode } from "./enums/AlphaMode";
 export { RenderQueueType } from "./enums/RenderQueueType";
 export { Material } from "./Material";
+export { BlinnPhongMaterial } from "./BlinnPhongMaterial";
 export { PBRMaterial } from "./PBRMaterial";
+export { PBRSpecularMaterial } from "./PBRSpecularMaterial";
+export { UnlitMaterial } from "./UnlitMaterial";
 
 // @todo: delete
 export type { TextureFormatDetail } from "./type";
