+++ packages/core/src/particle/ParticleRenderer.ts
@@ -1,4 +1,4 @@
-import { MathUtil, Vector3 } from "@oasis-engine/math";
+import { MathUtil, Vector3, Color } from "@oasis-engine/math";
 import { BufferGeometry, GeometryRenderer } from "../geometry";
 import { Buffer } from "../graphic/Buffer";
 import { BufferBindFlag } from "../graphic/enums/BufferBindFlag";
@@ -28,7 +28,15 @@ enum DirtyFlagType {
 }
 
 /**
- * Particle Renderer Component
+ * Blend mode enums of the particle renderer's material.
+ */
+export enum ParticleRendererBlendMode {
+  Transparent = 0,
+  Additive = 1
+}
+
+/**
+ * Particle Renderer Component.
  */
 export class ParticleRenderer extends GeometryRenderer {
   private static _getRandom(): number {
@@ -46,7 +54,7 @@ export class ParticleRenderer extends GeometryRenderer {
   private _velocityRandomness: Vector3 = new Vector3();
   private _acceleration: Vector3 = new Vector3();
   private _accelerationRandomness: Vector3 = new Vector3();
-  private _color: Vector3 = new Vector3(1, 1, 1);
+  private _color: Color = new Color(1, 1, 1, 1);
   private _colorRandomness: number = 0;
   private _size: number = 1;
   private _sizeRandomness: number = 0;
@@ -70,6 +78,8 @@ export class ParticleRenderer extends GeometryRenderer {
   private _is2d: boolean = true;
   private _isFadeIn: boolean = false;
   private _isFadeOut: boolean = false;
+  private _playOnEnable: boolean = true;
+  private _blendMode: ParticleRendererBlendMode = ParticleRendererBlendMode.Transparent;
 
   /**
    * Sprite sheet of texture.
@@ -179,11 +189,11 @@ export class ParticleRenderer extends GeometryRenderer {
   /**
    * Color of particles.
    */
-  get color(): Vector3 {
+  get color(): Color {
     return this._color;
   }
 
-  set color(value: Vector3) {
+  set color(value: Color) {
     this._updateDirtyFlag |= DirtyFlagType.Color;
     this._color = value;
   }
@@ -343,27 +353,30 @@ export class ParticleRenderer extends GeometryRenderer {
     this._isStart = false;
     this._isInit = false;
     this._maxCount = value;
+    this._updateDirtyFlag = DirtyFlagType.Everything;
     this.geometry = this._createGeometry();
 
     this._updateBuffer();
 
     this._isInit = true;
+
   }
 
   /**
-   * Play once or not.
+   * Whether play once.
    */
   get isOnce(): boolean {
     return this._isOnce;
   }
 
   set isOnce(value: boolean) {
+    this._time = 0;
     this.shaderData.setInt("u_once", value ? 1 : 0);
     this._isOnce = value;
   }
 
   /**
-   * Follow the direction of velocity or not.
+   * Whether follow the direction of velocity.
    */
   get isRotateToVelocity(): boolean {
     return this._isRotateToVelocity;
@@ -380,7 +393,7 @@ export class ParticleRenderer extends GeometryRenderer {
   }
 
   /**
-   * Follow the direction of velocity or not.
+   * Whether use origin color.
    */
   get isUseOriginColor(): boolean {
     return this._isUseOriginColor;
@@ -397,7 +410,7 @@ export class ParticleRenderer extends GeometryRenderer {
   }
 
   /**
-   * Is scale by lifetime  or not.
+   * Whether scale by lifetime.
    */
   get isScaleByLifetime(): boolean {
     return this._isScaleByLifetime;
@@ -414,7 +427,7 @@ export class ParticleRenderer extends GeometryRenderer {
   }
 
   /**
-   * Is 2D rendering or not.
+   * Whether 2D rendering.
    */
   get is2d(): boolean {
     return this._is2d;
@@ -425,13 +438,14 @@ export class ParticleRenderer extends GeometryRenderer {
       this.shaderData.enableMacro("is2d");
     } else {
       this.shaderData.disableMacro("is2d");
+      this.material.renderState.rasterState.cullMode = CullMode.Off;
     }
 
     this._is2d = value;
   }
 
   /**
-   * Is fade in or not.
+   * Whether fade in.
    */
   get isFadeIn(): boolean {
     return this._isFadeIn;
@@ -448,7 +462,7 @@ export class ParticleRenderer extends GeometryRenderer {
   }
 
   /**
-   * Is fade out or not.
+   * Whether fade out.
    */
   get isFadeOut(): boolean {
     return this._isFadeOut;
@@ -464,6 +478,47 @@ export class ParticleRenderer extends GeometryRenderer {
     this._isFadeOut = value;
   }
 
+  /**
+   * Whether play on enable.
+   */
+  get playOnEnable(): boolean {
+    return this._playOnEnable;
+  }
+
+  set playOnEnable(value: boolean) {
+    this._playOnEnable = value;
+
+    if (value) {
+      this.start();
+    }
+    else {
+      this.stop();
+    }
+  }
+
+  /**
+   * Blend mode of the particle renderer's material.
+   */
+  get blendMode(): ParticleRendererBlendMode {
+    return this._blendMode;
+  }
+
+  set blendMode(value: ParticleRendererBlendMode) {
+    const blendState = this.material.renderState.blendState;
+    const target = blendState.targetBlendState;
+
+    if (value === ParticleRendererBlendMode.Transparent) {
+      target.sourceColorBlendFactor = target.sourceAlphaBlendFactor = BlendFactor.SourceAlpha;
+      target.destinationColorBlendFactor = target.destinationAlphaBlendFactor = BlendFactor.OneMinusSourceAlpha;
+    }
+    else if (value === ParticleRendererBlendMode.Additive) {
+      target.sourceColorBlendFactor = target.sourceAlphaBlendFactor = BlendFactor.SourceAlpha;
+      target.destinationColorBlendFactor = target.destinationAlphaBlendFactor = BlendFactor.One;
+    }
+
+    this._blendMode = value;
+  }
+
   constructor(props) {
     super(props);
 
@@ -474,7 +529,7 @@ export class ParticleRenderer extends GeometryRenderer {
    * @override
    * @internal
    */
-  update(deltaTime: number) {
+  update(deltaTime: number): void {
     if (!this._isInit || !this._isStart) {
       return;
     }
@@ -489,22 +544,34 @@ export class ParticleRenderer extends GeometryRenderer {
   }
 
   /**
-   * start emitting
+   * @override
+   * @internal
    */
-  start() {
+  _onEnable(): void {
+    super._onEnable();
+
+    if (this._playOnEnable) {
+      this.start();
+    }
+  }
+
+  /**
+   * Start emitting.
+   */
+  start(): void {
     this._isStart = true;
     this._time = 0;
     this.shaderData.setInt("u_active", 1);
   }
 
   /**
-   * stop emitting
+   * Stop emitting.
    */
-  stop() {
+  stop(): void {
     this.shaderData.setInt("u_active", 0);
   }
 
-  private _createMaterial() {
+  private _createMaterial(): Material {
     const material = new Material(this.engine, Shader.find("particle-shader"));
     const { renderState } = material;
     const target = renderState.blendState.targetBlendState;
@@ -513,10 +580,6 @@ export class ParticleRenderer extends GeometryRenderer {
     target.destinationColorBlendFactor = target.destinationAlphaBlendFactor = BlendFactor.OneMinusSourceAlpha;
 
     renderState.depthState.writeEnabled = false;
-
-    if (!this.is2d) {
-      renderState.rasterState.cullMode = CullMode.Off;
-    }
 
     material.renderQueueType = RenderQueueType.Transparent;
 
@@ -527,7 +590,7 @@ export class ParticleRenderer extends GeometryRenderer {
     return material;
   }
 
-  private _createGeometry() {
+  private _createGeometry(): BufferGeometry {
     const geometry = new BufferGeometry(this._entity.engine, "particleGeometry");
     const vertexStride = 96;
     const vertexFloatCount = this._maxCount * 4 * vertexStride;
@@ -575,7 +638,7 @@ export class ParticleRenderer extends GeometryRenderer {
     return geometry;
   }
 
-  private _updateBuffer() {
+  private _updateBuffer(): void {
     for (let x = 0; x < this._maxCount; x++) {
       this._updateSingleBuffer(x);
     }
@@ -583,7 +646,7 @@ export class ParticleRenderer extends GeometryRenderer {
     this._vertexBuffer.setData(this._vertices);
   }
 
-  private _updateSingleBuffer(i: number) {
+  private _updateSingleBuffer(i: number): void {
     const { _updateDirtyFlag, _vertices: vertices, _vertexStride: vertexStride } = this;
     const { _getRandom: getRandom } = ParticleRenderer;
     const offset = i * 4;
@@ -643,17 +706,18 @@ export class ParticleRenderer extends GeometryRenderer {
       const { _color, _colorRandomness } = this;
 
       vertices[k0 + 9] = vertices[k1 + 9] = vertices[k2 + 9] = vertices[k3 + 9] = MathUtil.clamp(
-        _color.x + getRandom() * _colorRandomness,
+        _color.r + getRandom() * _colorRandomness,
         0,
         1
       );
+
       vertices[k0 + 10] = vertices[k1 + 10] = vertices[k2 + 10] = vertices[k3 + 10] = MathUtil.clamp(
-        _color.y + getRandom() * _colorRandomness,
+        _color.g + getRandom() * _colorRandomness,
         0,
         1
       );
       vertices[k0 + 11] = vertices[k1 + 11] = vertices[k2 + 11] = vertices[k3 + 11] = MathUtil.clamp(
-        _color.z + getRandom() * _colorRandomness,
+        _color.b + getRandom() * _colorRandomness,
         0,
         1
       );
@@ -705,7 +769,7 @@ export class ParticleRenderer extends GeometryRenderer {
     this._updateSingleUv(i, k0, k1, k2, k3);
   }
 
-  private _updateSingleUv(i: number, k0: number, k1: number, k2: number, k3: number) {
+  private _updateSingleUv(i: number, k0: number, k1: number, k2: number, k3: number): void {
     const { spriteSheet } = this;
     const texture = this._material.shaderData.getTexture("u_texture");
     const vertices = this._vertices;
@@ -747,22 +811,22 @@ export class ParticleRenderer extends GeometryRenderer {
 
         // left bottom
         vertices[k0 + 19] = 0;
-        vertices[k0 + 20] = 0;
+        vertices[k0 + 20] = 1;
         vertices[k0 + 21] = ratio;
 
         // right bottom
         vertices[k1 + 19] = 1;
-        vertices[k1 + 20] = 0;
+        vertices[k1 + 20] = 1;
         vertices[k1 + 21] = ratio;
 
         // right top
         vertices[k2 + 19] = 1;
-        vertices[k2 + 20] = 1;
+        vertices[k2 + 20] = 0;
         vertices[k2 + 21] = ratio;
 
         // left top
         vertices[k3 + 19] = 0;
-        vertices[k3 + 20] = 1;
+        vertices[k3 + 20] = 0;
         vertices[k3 + 21] = ratio;
       }
     } else {
