+++ packages/core/src/shaderlib/extra/unlit.vs.glsl
@@ -1,0 +1,13 @@
+#include <common_vert>
+#include <uv_share>
+#include <fog_share>
+
+void main() {
+
+    #include <begin_position_vert>
+    #include <skinning_vert>
+    #include <uv_vert>
+    #include <position_vert>
+
+    #include <fog_vert>
+}
