+++ packages/core/src/shaderlib/skinning_vert.glsl
@@ -17,7 +17,7 @@
 
         position = skinMatrix * position;
 
-        #ifdef O3_HAS_NORMAL
+        #if defined(O3_HAS_NORMAL) && !defined(OMIT_NORMAL)
             normal = vec4( skinMatrix * vec4( normal, 0.0 ) ).xyz;
             #if defined( O3_HAS_TANGENT ) && defined( O3_HAS_NORMALMAP )
                 tangent.xyz = vec4( skinMatrix * vec4( tangent.xyz, 0.0 ) ).xyz;
