+++ packages/design/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oasis-engine/design",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "main": "dist/main.js",
   "module": "dist/module.js",
