+++ packages/draco/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oasis-engine/draco",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "scripts": {
     "b:types": "tsc"
@@ -13,6 +13,6 @@
     "types/**/*"
   ],
   "dependencies": {
-    "@oasis-engine/core": "0.2.0"
+    "@oasis-engine/core": "0.2.1"
   }
 }
