+++ packages/loader/src/scene-loader/resources/index.ts
@@ -1,6 +1,8 @@
 export { SchemaResource } from "./SchemaResource";
 export { GLTFResource } from "./GLTFResource";
 export { PBRMaterialResource } from "./PBRMaterialResource";
+export { PBRSpecularMaterialResource } from "./PBRSpecularMaterialResource";
+export { UnlitMaterialResource } from "./UnlitMaterialResource";
 export { TextureResource } from "./TextureResource";
 export { ScriptResource } from "./ScriptResource";
 export { BlinnPhongMaterialResource } from "./BlinnPhongMaterialResource";
