+++ packages/oasis-engine/package.json
@@ -1,6 +1,6 @@
 {
   "name": "oasis-engine",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "scripts": {
     "b:types": "tsc"
@@ -14,9 +14,9 @@
     "types/**/*"
   ],
   "dependencies": {
-    "@oasis-engine/core": "0.2.0",
-    "@oasis-engine/loader": "0.2.0",
-    "@oasis-engine/math": "0.2.0",
-    "@oasis-engine/rhi-webgl": "0.2.0"
+    "@oasis-engine/core": "0.2.1",
+    "@oasis-engine/loader": "0.2.1",
+    "@oasis-engine/math": "0.2.1",
+    "@oasis-engine/rhi-webgl": "0.2.1"
   }
 }
