+++ packages/stats/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oasis-engine/stats",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "license": "MIT",
   "scripts": {
     "b:types": "tsc"
@@ -14,6 +14,6 @@
     "types/**/*"
   ],
   "dependencies": {
-    "oasis-engine": "0.2.0"
+    "oasis-engine": "0.2.1"
   }
 }
